pyproject.toml
[project]
name = "parking-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.5.3",
    "coverage>=7.10.6",
    "django>=5.2.4",
    "django-jazzmin>=3.0.1",
    "django-ninja>=1.4.3",
    "django-ninja-jwt>=5.3.7",
    "playwright>=1.55.0",
    "psycopg2-binary>=2.9.10",
]


[tool.ruff]
line-length = 79
exclude = ["__init__.py", 'apps.py']

[tool.ruff.lint]
select = ["E", "F", "I", "W", "PL", "PT", "D", "ANN"]


[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
disallow_untyped_defs = true
warn_unused_ignores = true
show_error_codes = true

plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "core.settings.settings"

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=5.2.2",
]
